pyproject.toml
[tool.poetry]
name = "nlp-server"
version = "0.2.0"
description = "Originated from spacy-cookiecutter by Microsoft"
authors = ["Daniel Chalef <131175+danielchalef@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
spacy = "^3.6.0"
uvicorn = "^0.23.1"
fastapi = "^0.100.0"
sentence-transformers = "^2.2.2"
orjson = "^3.9.2"
pydantic-numpy = "^2.2.2"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.3.0"
ruff = "^0.0.269"
black = "^23.3.0"
pytest = "^7.3.1"
locust = "^2.15.1"
httpx = "^0.24.1"
types-pyyaml = "^6.0.12.11"

[[tool.poetry.dependencies.torch]]
version = "^2.0.1"
markers = "platform_machine == 'x86_64'"
source = "torch"

[[tool.poetry.dependencies.torch]]
version = "^2.0.1"
markers = "platform_machine != 'x86_64'"
source = "PyPI"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["pydantic.mypy"]
